requirements.txt
Django==4.2.17
Pillow==11.0.0
gunicorn
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
django-storages
psycopg2-binary
